+++ yastn/tensor/__init__.py
@@ -103,7 +103,7 @@ class Tensor:
                     raise YastnError("Diagonal tensor should have s equal (1, -1) or (-1, 1)")
                 if any(x != 0 for x in n):
                     raise YastnError("Tensor charge of a diagonal tensor should be 0")
-            self.struct = _struct(s=s, n=n, diag=isdiag)
+            self.struct = _struct(s=s, n=n, diag=bool(isdiag))
 
         self.slices = kwargs['slices'] if 'slices' in kwargs else ()
 
@@ -127,9 +127,9 @@ class Tensor:
     from ._single import conj, conj_blocks, flip_signature, flip_charges, transpose, moveaxis, move_leg, diag, grad
     from ._single import copy, clone, detach, to, requires_grad_, remove_zero_blocks, add_leg, remove_leg, drop_leg_history
     from ._output import show_properties, __str__, print_blocks_shape, is_complex
-    from ._output import get_blocks_charge, get_blocks_shape, get_leg_charges_and_dims, get_leg_structure, get_legs
+    from ._output import get_blocks_charge, get_blocks_shape, get_legs
     from ._output import zero_of_dtype, item, __getitem__, __contains__
-    from ._output import get_leg_fusion, get_shape, get_signature, get_dtype
+    from ._output import get_shape, get_signature, get_dtype
     from ._output import get_tensor_charge, get_rank
     from ._output import to_number, to_dense, to_numpy, to_raw_tensor, to_nonsymmetric
     from ._output import save_to_hdf5, save_to_dict, compress_to_1d