+++ CHANGELOG.md
@@ -212,3 +212,11 @@ v0.9
 11-2023
 - Tensor.transpose() and Tensor.T reverse th order of axes
 - make_config  accepts str for predefined symmetries
+
+06-2024
+- tag v1.0
+
+07-2024  (v1.01)
+- fix engine type consistency for numpy2.0
+- consistently use 'U1' for U(1)-symmetry
+- allow using **config._asdict() while initializing operator classes